+++ bottomsheet/src/main/java/com/trendyol/design/bottomsheet/preview/PreviewTheme.kt
@@ -1,0 +1,46 @@
+package com.trendyol.design.bottomsheet.preview
+
+import androidx.compose.material.Colors
+import androidx.compose.material.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.ReadOnlyComposable
+import androidx.compose.ui.graphics.Color
+import com.trendyol.theme.LocalTrendyolDesignTheme
+import com.trendyol.theme.TrendyolDesignColors
+import com.trendyol.theme.TrendyolDesignFontFamily
+import com.trendyol.theme.TrendyolDesignTheme
+import com.trendyol.theme.TrendyolDesignTypography
+
+@Composable
+internal fun PreviewTheme(
+    content: @Composable () -> Unit
+) {
+    MaterialTheme(
+        colors = TrendyolColor.materialColor
+    ) {
+        CompositionLocalProvider(
+            LocalTrendyolDesignTheme provides trendyolDesignTheme
+        ) {
+            content()
+        }
+    }
+}
+
+private object TrendyolColor {
+    val themeColor = TrendyolDesignColors().copy(colorPrimary = colorPrimary)
+    val materialColor: Colors
+        @ReadOnlyComposable
+        @Composable
+        get() = MaterialTheme.colors.copy(
+            primary = colorPrimary
+        )
+}
+
+private val colorPrimary = Color(color = 0xFFF27A1A)
+
+private val trendyolDesignTheme = TrendyolDesignTheme(
+    colors = TrendyolDesignColors(),
+    typography = TrendyolDesignTypography(),
+    fontFamily = TrendyolDesignFontFamily(),
+)